+++ Extensions/EntityQueryExtensions.cs
@@ -1,3 +1,4 @@
+using Ardalis.GuardClauses;
 using Serenity.Data;
 
 namespace Idevs.Extensions;
@@ -12,8 +13,7 @@ public static class EntityQueryExtensions
     ///   Object itself.</returns>
     public static T Set<T>(this T self, IRow row, params Field[] exclude) where T : ISetFieldByStatement
     {
-        if (row == null)
-            throw new ArgumentNullException("row");
+        Guard.Against.Null(row, nameof(row));
 
         if (!row.TrackAssignments)
             throw new ArgumentException("row must be in TrackAssignments mode to determine modified fields.");