+++ Extensions/TextLocalizerExtensions.cs
@@ -10,7 +10,9 @@ public static class TextLocalizerExtension
         return localizer.TryGet(name) ?? key;
     }
 
-    public static string TranslateText(this string key, string moduleName, ITextLocalizer localizer) => localizer?.TryGet($"{moduleName}.{key}") ?? key;
+    public static string TranslateText(this string key, string moduleName, ITextLocalizer? localizer) =>
+        localizer?.TryGet($"{moduleName}.{key}") ?? key;
 
-    public static string TranslateData(this string key, string moduleName, ITextLocalizer localizer) => localizer.Translate(moduleName, key);
+    public static string TranslateData(this string key, string moduleName, ITextLocalizer localizer) =>
+        localizer.Translate(moduleName, key);
 }