+++ Repositories/RepositoryBase.cs
@@ -9,7 +9,11 @@ public class RepositoryBase
 {
     protected ISqlConnections SqlConnections { get; }
     protected IDbConnection Connection => SqlConnections.NewByKey("Default");
+#if NET6_0
     protected IExceptionLogger ExceptionLog { get; }
+#else
+    protected Microsoft.Extensions.Logging.ILogger ExceptionLog {get;}
+#endif
     protected ITextLocalizer Localizer { get; }
 
     protected SqlQuery SqlQuery => new SqlQuery().Dialect(SqlServer2012Dialect.Instance);
@@ -17,6 +21,7 @@ public class RepositoryBase
     protected SqlUpdate SqlUpdate(string tableName) => new SqlUpdate(tableName).Dialect(SqlServer2012Dialect.Instance);
     protected SqlDelete SqlDelete(string tableName) => new SqlDelete(tableName);
 
+#if NET6_0
     public RepositoryBase(ISqlConnections sqlConnections, IExceptionLogger exceptionLogger, ITextLocalizer localizer)
     {
         SqlConnections = sqlConnections
@@ -25,4 +30,14 @@ public class RepositoryBase
                        ?? throw new ArgumentNullException(nameof(exceptionLogger));
         Localizer = localizer ?? throw new ArgumentNullException(nameof(localizer));
     }
+#else
+    public RepositoryBase(ISqlConnections sqlConnections, Microsoft.Extensions.Logging.ILogger exceptionLogger, ITextLocalizer localizer)
+    {
+        SqlConnections = sqlConnections
+                         ?? throw new ArgumentNullException(nameof(sqlConnections));
+        ExceptionLog = exceptionLogger
+                       ?? throw new ArgumentNullException(nameof(exceptionLogger));
+        Localizer = localizer ?? throw new ArgumentNullException(nameof(localizer));
+    }
+#endif
 }